% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funding.R
\name{funding_general}
\alias{funding_general}
\alias{funding_catalyst}
\alias{funding_cfar}
\alias{funding_hsci}
\alias{funding_tnt}
\alias{funding_niehs}
\title{Return acknowledgement section for analysis done by HBC}
\usage{
funding_general(author, service)

funding_catalyst()

funding_cfar(author)

funding_hsci(author)

funding_tnt(author)

funding_niehs(author)
}
\arguments{
\item{author}{Name of the author/s to be added to the text.}

\item{service}{Name of the analysis done by the core.}
}
\description{
Return acknowledgement section for analysis done by HBC
}
\examples{
funding_general("Lorena Pantano", "RNASeq...")
funding_catalyst()
}
