% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{show_scripts}
\alias{show_scripts}
\title{Show all available scripts}
\usage{
show_scripts(category = "singlecell")
}
\arguments{
\item{category}{Name of type of analysis. (singlecell for now)}
}
\description{
Show all available scripts
}
