% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{copy_script}
\alias{copy_script}
\title{Copy a given script to other place}
\usage{
copy_script(category, fn, dest = NULL, ...)
}
\arguments{
\item{category}{Name of type of analysis.}

\item{fn}{Name of the file to copy.}
}
\description{
Copy a given script to other place
}
